Rename the cookie banner's action element to action-button
`<govuk-cookie-banner-message-action>` becomes
`<govuk-cookie-banner-message-action-button>`, matching the panel
component's `<govuk-panel-action-button>` and pairing properly with the
`<govuk-cookie-banner-message-action-link>` that sits next to it.

The old element keeps working — it is now a deprecated subclass of the
renamed tag helper, so it generates exactly what it did before — but
using it produces a deprecation warning with the diagnostic ID GFA0006.
`HtmlTargetElementAttribute` is not inherited, so the subclass matches
only the old tag name.

Fixes #488
