+++ src/GovUk.Frontend.AspNetCore.Docs/Templates/components/cookie-banner.liquid
@@ -23,6 +23,6 @@
 
 {{ tag_helper_api("CookieBannerMessageActionsTagHelper") }}
 
-{{ tag_helper_api("CookieBannerMessageActionTagHelper") }}
+{{ tag_helper_api("CookieBannerMessageActionButtonTagHelper") }}
 
 {{ tag_helper_api("CookieBannerMessageActionLinkTagHelper") }}
