+++ src/GovUk.Frontend.AspNetCore/TagHelpers/CookieBannerMessageActionTagHelper.cs
@@ -1,75 +1,20 @@
-using GovUk.Frontend.AspNetCore.ComponentGeneration;
+using System.ComponentModel;
 using Microsoft.AspNetCore.Razor.TagHelpers;
 
 namespace GovUk.Frontend.AspNetCore.TagHelpers;
 
 /// <summary>
-/// Represents an action in the message in a GDS cookie banner component.
+/// Represents a button action in the message in a GDS cookie banner component.
 /// </summary>
+/// <remarks>
+/// This element has been replaced by <c>govuk-cookie-banner-message-action-button</c>.
+/// </remarks>
 [HtmlTargetElement(TagName, ParentTag = CookieBannerMessageActionsTagHelper.TagName, TagStructure = TagStructure.WithoutEndTag)]
-public class CookieBannerMessageActionTagHelper : TagHelper
+[EditorBrowsable(EditorBrowsableState.Never)]
+[Obsolete(
+    "Use the <" + CookieBannerMessageActionButtonTagHelper.TagName + "> element instead.",
+    DiagnosticId = DiagnosticIds.UseCookieBannerMessageActionButtonElementInstead)]
+public class CookieBannerMessageActionTagHelper : CookieBannerMessageActionButtonTagHelper
 {
-    internal const string TagName = "govuk-cookie-banner-message-action";
-
-    private const string NameAttributeName = "name";
-    private const string TextAttributeName = "text";
-    private const string TypeAttributeName = "type";
-    private const string ValueAttributeName = "value";
-
-    /// <summary>
-    /// The <c>name</c> attribute for the generated <c>button</c> element.
-    /// </summary>
-    [HtmlAttributeName(NameAttributeName)]
-    public string? Name { get; set; }
-
-    /// <summary>
-    /// The button text.
-    /// </summary>
-    /// <remarks>
-    /// This attribute is required.
-    /// </remarks>
-    [HtmlAttributeName(TextAttributeName)]
-    public string? Text { get; set; }
-
-    /// <summary>
-    /// The <c>type</c> attribute for the generated <c>button</c> element.
-    /// </summary>
-    [HtmlAttributeName(TypeAttributeName)]
-    public string? Type { get; set; }
-
-    /// <summary>
-    /// The <c>value</c> attribute for the generated <c>button</c> element.
-    /// </summary>
-    [HtmlAttributeName(ValueAttributeName)]
-    public string? Value { get; set; }
-
-    /// <inheritdoc/>
-    public override void Process(TagHelperContext context, TagHelperOutput output)
-    {
-        ArgumentNullException.ThrowIfNull(context);
-        ArgumentNullException.ThrowIfNull(output);
-
-        var actionsContext = context.GetContextItem<CookieBannerMessageActionsContext>();
-
-        if (Text is null)
-        {
-            throw ExceptionHelper.TheAttributeMustBeSpecified(TextAttributeName);
-        }
-
-        var attributes = new AttributeCollection(output.Attributes);
-        attributes.Remove("class", out var classes);
-
-        actionsContext.Actions.Add(new CookieBannerOptionsMessageAction
-        {
-            Text = Text,
-            Type = Type,
-            Href = null,
-            Name = Name,
-            Value = Value,
-            Classes = classes,
-            Attributes = attributes
-        });
-
-        output.SuppressOutput();
-    }
+    internal new const string TagName = "govuk-cookie-banner-message-action";
 }
