+++ tests/GovUk.Frontend.AspNetCore.Tests/TagHelpers/CookieBannerMessageActionTagHelperTests.cs
@@ -1,39 +1,29 @@
 using GovUk.Frontend.AspNetCore.TagHelpers;
+#pragma warning disable GFA0006 // Type or member is obsolete
 
 namespace GovUk.Frontend.AspNetCore.Tests.TagHelpers;
 
 public class CookieBannerMessageActionTagHelperTests : TagHelperTestBase<CookieBannerMessageActionTagHelper>
 {
     [Fact]
-    public async Task ProcessAsync_AddsActionToContext()
+    public async Task ProcessAsync_DeprecatedTagName_AddsActionToContext()
     {
         // Arrange
         var text = "Action";
-        var name = "Name";
         var type = "type";
-        var value = "Value";
-        var className = CreateDummyClassName();
-        var attributes = CreateDummyDataAttributes();
 
         var actionsContext = new CookieBannerMessageActionsContext();
         var messageContext = new CookieBannerMessageContext() { Actions = actionsContext };
         var cookieBannerContext = new CookieBannerContext();
 
-        var context = CreateTagHelperContext(
-            className: className,
-            attributes: attributes,
-            contexts: [cookieBannerContext, messageContext, actionsContext]);
+        var context = CreateTagHelperContext(contexts: [cookieBannerContext, messageContext, actionsContext]);
 
-        var output = CreateTagHelperOutput(
-            className: className,
-            attributes: attributes);
+        var output = CreateTagHelperOutput();
 
         var tagHelper = new CookieBannerMessageActionTagHelper()
         {
-            Name = name,
             Text = text,
-            Type = type,
-            Value = value
+            Type = type
         };
 
         tagHelper.Init(context);
@@ -47,12 +37,8 @@ public class CookieBannerMessageActionTagHelperTests : TagHelperTestBase<CookieB
             action =>
             {
                 Assert.Equal(text, action.Text);
-                Assert.Equal(className, action.Classes);
-                Assert.Null(action.Href);
-                Assert.Equal(name, action.Name);
-                Assert.Equal(value, action.Value);
                 Assert.Equal(type, action.Type);
-                AssertContainsAttributes(attributes, action.Attributes);
+                Assert.Null(action.Href);
             });
     }
 }
